// src/app/page/master/map-jenispetugas-jenispegawai/map-jenispetugas-jenispegawai.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapJenispetugasJenispegawai {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}